refactor: 개별 발송, 그룹 발송 기능을 하나의 페이지에서 가능하도록 변경
