+++ src/main/kotlin/apply/ui/admin/BaseLayout.kt
@@ -2,7 +2,7 @@ package apply.ui.admin
 
 import apply.ui.admin.cheater.CheatersView
 import apply.ui.admin.evaluation.EvaluationsView
-import apply.ui.admin.mail.MailSelectionsView
+import apply.ui.admin.mail.MailFormView
 import apply.ui.admin.recruitment.RecruitmentsView
 import apply.ui.admin.selections.SelectionsView
 import com.vaadin.flow.component.Component
@@ -27,7 +27,7 @@ class BaseLayout : AppLayout() {
         "평가 관리" to EvaluationsView::class.java,
         "선발 과정" to SelectionsView::class.java,
         "부정 행위자" to CheatersView::class.java,
-        "메일 발송" to MailSelectionsView::class.java
+        "메일 발송" to MailFormView::class.java
     )
 
     init {
